refactor: stimulusファイル等
