+++ app/views/shared/_header.html.erb
@@ -2,8 +2,7 @@
 <header class="bg-white dark:bg-black
                border-b border-gray-200 dark:border-gray-800
                sticky top-0 z-50
-               transition-colors duration-200"
-        data-controller="mobile-menu">
+               transition-colors duration-200">
   <div class="container mx-auto px-4 max-w-7xl">
     <div class="flex items-center justify-between h-14">
